+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/peripheral/IPeripheralOwnerAbility.kt
@@ -1,0 +1,3 @@
+package site.siredvin.peripheralium.api.peripheral
+
+interface IPeripheralOwnerAbility<T : IOwnerAbility>